requirements.txt
Django
psycopg2-binary
djangorestframework
django-admin-interface
django-colorfield
django-environ
whitenoise
celery
redis
django[argon2]
argon2-cffi